+++ src/mympd_api/mympd_api_handler.c
@@ -37,6 +37,7 @@
 #include "src/mympd_api/outputs.h"
 #include "src/mympd_api/partitions.h"
 #include "src/mympd_api/pictures.h"
+#include "src/mympd_api/playlistart.h"
 #include "src/mympd_api/playlists.h"
 #include "src/mympd_api/queue.h"
 #include "src/mympd_api/requests.h"
@@ -245,12 +246,12 @@ void mympd_api_handler(struct t_mympd_state *mympd_state, struct t_partition_sta
         case MYMPD_API_STATS:
             response->data = mympd_api_stats_get(partition_state, response->data, request->id);
             break;
-    // Tagart
-        case INTERNAL_API_TAGART:
-            if (json_get_string(request->data, "$.params.tag", 1, NAME_LEN_MAX, &sds_buf1, vcb_ismpdtag, &parse_error) == true &&
-                json_get_string(request->data, "$.params.value", 1, NAME_LEN_MAX, &sds_buf2, vcb_isname, &parse_error) == true)
+    // Playlistart
+        case INTERNAL_API_PLAYLISTART:
+            if (json_get_string(request->data, "$.params.name", 1, NAME_LEN_MAX, &sds_buf1, vcb_isname, &parse_error) == true &&
+                json_get_string(request->data, "$.params.type", 1, NAME_LEN_MAX, &sds_buf2, vcb_isname, &parse_error) == true)
             {
-                response->data = mympd_api_tagart(mympd_state, partition_state, response->data, request->id, request->conn_id, sds_buf1, sds_buf2);
+                response->data = mympd_api_playlistart(mympd_state, partition_state, response->data, request->id, request->conn_id, sds_buf1, sds_buf2);
                 if (sdslen(response->data) == 0) {
                     // response must be send by triggered script
                     async = true;
@@ -259,6 +260,20 @@ void mympd_api_handler(struct t_mympd_state *mympd_state, struct t_partition_sta
                 }
             }
             break;
+    // Tagart
+    case INTERNAL_API_TAGART:
+    if (json_get_string(request->data, "$.params.tag", 1, NAME_LEN_MAX, &sds_buf1, vcb_ismpdtag, &parse_error) == true &&
+        json_get_string(request->data, "$.params.value", 1, NAME_LEN_MAX, &sds_buf2, vcb_isname, &parse_error) == true)
+    {
+        response->data = mympd_api_tagart(mympd_state, partition_state, response->data, request->id, request->conn_id, sds_buf1, sds_buf2);
+        if (sdslen(response->data) == 0) {
+            // response must be send by triggered script
+            async = true;
+            // we do not pass the request to the script thread
+            free_request(request);
+        }
+    }
+    break;
     // Albumart
         case INTERNAL_API_ALBUMART_BY_URI:
             if (json_get_string(request->data, "$.params.uri", 1, FILEPATH_LEN_MAX, &sds_buf1, vcb_isfilepath, &parse_error) == true) {