+++ src/mympd_api/trigger.h
@@ -33,6 +33,7 @@ enum trigger_events {
     TRIGGER_MYMPD_JUKEBOX = -11,       //!< myMPD jukebox
     TRIGGER_MYMPD_SMARTPLS = - 12,     //!< myMPD smart playlist creation
     TRIGGER_MYMPD_BGIMAGE = -13,       //!< myMPD background image
+    TRIGGER_MYMPD_PLAYLISTART = -14,   //!< myMPD playlistart
     TRIGGER_MPD_DATABASE = 0x1,        //!< mpd database has changed
     TRIGGER_MPD_STORED_PLAYLIST = 0x2, //!< mpd playlist idle event
     TRIGGER_MPD_QUEUE = 0x4,           //!< mpd queue idle event